Updated schedule extractor to use util function + M25 support
