+++ scheduleExtractor/buildExe.bat
@@ -1,0 +1,1 @@
+nexe --build -i scheduleExtractor.js -t x64-14.15.3 -r "../node_modules/madden-franchise/data/schemas" -r "../Utils/FranchiseUtils.js" -r "../Utils/ScheduleFunctions.js" -r "../Utils/FranchiseTableId.js" -r "../retroSchedules/teamLookup.json" -o "scheduleExtractor.exe" --verbose
