--- scheduleExtractorM24/buildExe.bat
@@ -1,1 +1,0 @@
-nexe --build -i scheduleExtractorM24.js -t x64-14.15.3 -r "../node_modules/madden-franchise/data/schemas" -r "../Utils/FranchiseUtils.js" -r "../Utils/FranchiseTableId.js" -r *.json -o "scheduleExtractorM24.exe" --verbose
